MPDControlPlugin: Move MPDClient wrapper into dedicated module
